requirements.txt
aiofiles==0.5.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
asyncio==3.4.3
attrs==19.3.0
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.1.7
biopython==1.77
bleach==3.1.5
blinker==1.4
Brlapi==0.7.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
cmake==3.18.0
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.8
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.11.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.7
neo4j==4.0.1
neo4j-driver==1.7.2
neobolt==1.7.17
neotime==1.7.4
netifaces==0.10.4
notebook==6.1.3
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
powerline-status==2.7
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.3
pymacaroons==0.13.0
pymol==2.3.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
ranger-fm==1.9.3
readline==6.2.4.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
sip==4.19.21
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
systemd-python==234
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.25
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xkit==0.0.0
